Updating numpy

+++ safe_control_gym/controllers/ppo/ppo_utils.py
@@ -235,7 +235,7 @@ class MLPActorCritic(nn.Module):
             ):
         dist, _ = self.actor(obs)
         action = dist.mode()
-        return action.cpu().numpy()
+        return action.cpu().numpy().astype(np.float32)
 
 
 class PPOBuffer(object):
@@ -303,7 +303,7 @@ class PPOBuffer(object):
             vshape = info['vshape']
             dtype = info.get('dtype', np.float32)
             init = info.get('init', np.zeros)
-            self.__dict__[k] = init(vshape, dtype=dtype)
+            self.__dict__[k] = init(vshape).astype(dtype)
         self.t = 0
 
     def push(self,
